@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
-require_relative "version"
-require_relative "uta"
+require_relative 'version'
+require_relative 'uta'
 module Utanone
   class ParseError < StandardError; end
 end
